requirements.txt
# Core UI Framework
streamlit>=1.24.0

# LLM + Google Gemini
google-generativeai>=0.3.2

# NLP & Evaluation
nltk>=3.8.1
rouge-score>=0.1.2
sentencepiece>=0.1.99

# File Handling
PyPDF2>=3.0.1
python-docx>=1.0.1
# textract removed due to version conflict
pandas>=2.0.0
numpy>=1.24.0

# Compatibility
protobuf>=3.20.0
